refactor: dedup checkpoint types + extract pure mission module (Tranche A)
Safe, behavior-frozen plumbing ahead of the checkpoint-system unification:

- New src/lib/mission/missions.ts: MISSION_GOALS, MISSION_CHECKPOINTS,
  checkMissionGoal, CheckpointType, and the canonical MissionCheckpoint type,
  moved verbatim out of MissionController.tsx (data unchanged).
- Merge the two duplicate row types (store's MissionCheckpoint + MissionController's
  Checkpoint) into ONE canonical MissionCheckpoint: field superset, with the
  tighter visual?: 'traffic-light' literal and the load-bearing orientation? /
  yawTolerance? fields (RoadProps reads orientation; mirror grading uses targetYaw).
- store.ts re-exports MissionCheckpoint so scoring/checkpointEval/useRegisterCheckpoint
  importers are unchanged.
- Repoint store/RoadProps/Scene/useMission to the pure module. This breaks the
  lib->component import cycle (store no longer imports the MissionController component).

Zero behavior change: type-check 0, 47 unit tests pass UNCHANGED, lint 0 errors,
next build OK. (The 2 scoring-test traffic-light fixtures gained `as const` to
satisfy the tightened visual literal — same values.)

Tranche B (make MISSION_CHECKPOINTS the single source for grading+scoring, delete
the dynamic activeCheckpoints registry; scores preserved) follows in its own PR.
